requirements.txt
colorlog
PyPDF2
ebooklib
python-docx
beautifulsoup4
